Move helper structs from view.go to util.go
